+++ Source/Artifacts/piggybank.js
@@ -1,0 +1,4 @@
+const Artifact = require("../../Classes/Artifact.js");
+
+module.exports = new Artifact("Piggy Bank", "Gain @{copies*5}% of your current gold when you enter a new room.")
+	.setElement("Untyped")